Move modules to core dir
