+++ shared/core/database/build.gradle.kts
@@ -2,7 +2,9 @@ import util.libs
 
 plugins {
     `kmm-domain-plugin`
+    kotlin("kapt")
     id("com.squareup.sqldelight")
+    id("dagger.hilt.android.plugin")
 }
 
 android {
@@ -14,6 +16,8 @@ dependencies {
     commonMainImplementation(libs.squareup.sqldelight.runtime)
 
     androidMainImplementation(libs.squareup.sqldelight.driver.android)
+    androidMainImplementation(libs.hilt.android)
+    kapt(libs.hilt.compiler)
 
     iosMainImplementation(libs.koin.core)
     iosMainImplementation(libs.squareup.sqldelight.driver.native)
