+++ shared/core/remote/build.gradle.kts
@@ -2,12 +2,14 @@
 import com.codingfeline.buildkonfig.compiler.FieldSpec.Type.STRING
 import util.libs
 import java.io.FileInputStream
-import java.util.Properties
+import java.util.*
 
 plugins {
     `kmm-domain-plugin`
     kotlin("plugin.serialization") version ("1.6.10")
+    kotlin("kapt")
     id("com.codingfeline.buildkonfig")
+    id("dagger.hilt.android.plugin")
 }
 
 android {
@@ -21,13 +23,17 @@ dependencies {
     commonMainImplementation(libs.koin.core)
     commonMainImplementation(libs.kermit)
 
+    androidMainImplementation(project(":shared:core:util"))
+
     androidMainImplementation(libs.ktor.android)
     androidMainImplementation(libs.squareup.sqldelight.driver.android)
+    androidMainImplementation(libs.hilt.android)
+    kapt(libs.hilt.compiler)
 
     iosMainImplementation(libs.ktor.ios)
 
     commonTestImplementation(kotlin("test"))
-    commonTestImplementation(projects.shared.core.test)
+    commonTestImplementation(project(":shared:core:test"))
 
     commonTestImplementation(libs.testing.ktor.mock)
     commonTestImplementation(libs.testing.turbine)
